% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3batchmark-package}
\alias{mlr3batchmark}
\alias{mlr3batchmark-package}
\title{mlr3batchmark: Batch Experiments for 'mlr3'}
\description{
Extends the 'mlr3' package with a connector to the package 'batchtools'. This allows to run large-scale benchmark experiments on scheduled high-performance computing clusters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https:///mlr3batchmark.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3batchmark}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3batchmark/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
}

Other contributors:
\itemize{
  \item Toby Hocking (\href{https://orcid.org/0000-0002-3146-0865}{ORCID}) [contributor]
}

}
