% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceResultsBatchmark.R
\name{reduceResultsBatchmark}
\alias{reduceResultsBatchmark}
\title{Collect Results from batchmark}
\usage{
reduceResultsBatchmark(ids = NULL, reg = batchtools::getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to the return value of \code{\link[batchtools]{findDone}}.
Invalid ids are ignored.}

\item{reg}{[\code{\link[batchtools]{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link[batchtools]{setDefaultRegistry}}).}
}
\value{
\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}.
}
\description{
Collect the results from jobs defined via \code{\link[=batchmark]{batchmark()}} and combine them into
a \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}.
}
